third_party/blink/web_tests/platform/mac-mac11-arm64/virtual/scroll-unification/plugins/focus-change-1-no-change-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(false)
i1focus
i1blur
pfocus
pblur
i2focus
PASS successfullyParsed is true

TEST COMPLETE


Test no change of focus inside plugin onfocus and onblur events

  
